update tests because of changes in how async node module handles async function with respect to mocha test execution, also because mocha now does not exit if something keeps the process open

+++ api_server/lib/test_base/codestream_api_test.js
@@ -114,14 +114,19 @@ class CodeStreamAPITest extends APIRequestTest {
 			this.mockMode &&
 			!this.testDidNotRun
 		) {
-			this.clearMockCache(callback);
+			this.clearMockCache(() => {
+				super.after(callback);
+			});
 		}
 		else {
-			callback();
+			super.after(callback);
 		}
 	}
 
 	clearMockCache (callback) {
+		if (!this.connectedToIpc()) { 
+			return callback();
+		}
 		this.doApiRequest(
 			{
 				method: 'delete',
